refactor: implement semantic playbook standards for type system
Add comprehensive semantic playbook and migrate all response types to
standardized DTO structure following NestJS best practices.

## Changes

### New Standards Document
- Add SEMANTIC_PLAYBOOK.md with comprehensive type system standards
- Define clear patterns for DTOs, responses, and naming conventions
- Document @SdkContract decorator usage and security patterns
- Add enum handling guidelines for contract extraction

### Type System Standardization
- Move all response types from api-responses.ts to module DTOs as classes
- Create 18 new response DTO files across 7 modules
- Convert all response interfaces to classes for consistency
- Standardize naming: all inputs end with 'Dto', all responses end with 'Response'

### Module Organization
**API Keys Module**
- Add api-key-response.dto.ts (ApiKeyResponse, ApiKeyListResponse, ApiKeyRollResponse)

**Auth Module**
- Add permission-response.dto.ts (PermissionResponse)
- Add @SdkContract to GET /whoami endpoint

**Messages Module**
- Add message-list-response.dto.ts (MessageListResponse)
- Add message-stats-response.dto.ts (MessageStatsResponse)
- Add received-message-response.dto.ts (ReceivedMessageResponse)
- Add sent-message-response.dto.ts (SentMessageResponse)

**Platforms Module**
- Add platform-response.dto.ts (PlatformResponse)
- Add supported-platforms-response.dto.ts (SupportedPlatformsResponse)
- Add platform-logs-response.dto.ts (PlatformLogsResponse, PlatformLogResponse)
- Add platform-log-stats-response.dto.ts (PlatformLogStatsResponse)
- Add query-platform-logs.dto.ts (QueryPlatformLogsDto)
- Extract QueryPlatformLogsDto from controller to proper DTO file

**Platform Messages Module**
- Add message-send-response.dto.ts (MessageSendResponse)
- Add message-status-response.dto.ts (MessageStatusResponse)
- Add message-retry-response.dto.ts (MessageRetryResponse)

**Projects Module**
- Add project-response.dto.ts (ProjectResponse)

**Users Module**
- Add add-member.dto.ts (AddMemberDto)
- Add update-member-role.dto.ts (UpdateMemberRoleDto)
- Add project-member-response.dto.ts (ProjectMemberResponse)

### Enum Handling
- Add src/common/types/enums.ts for Prisma enum re-exports
- Re-export ProjectRole and ProjectEnvironment enums
- Update DTOs to import enums from common/types/enums.ts instead of @prisma/client
- Enable contract extractor to auto-discover enum types

### Cleanup
- Remove src/common/types/sdk-models.ts (obsolete, completely unused)
- Clean api-responses.ts to keep only MessageResponse (generic type)
- Update all controller imports to use module DTOs
- Update 9 controllers with proper DTO references

## Breaking Changes

Response type names now have 'Response' suffix:
- Identity → IdentityResponse
- Project → ProjectResponse
- etc.

SDK/CLI regeneration required after merge.

## Verification

- ✅ Build: Clean compilation
- ✅ Tests: 656/656 passing
- ✅ Contract extraction: 50 contracts, 62 types
- ✅ All DTOs follow naming conventions (Dto/Response suffixes)
- ✅ No type definitions remain in controllers
- ✅ Zero unused files or imports
