+++ src/messages/dto/message-list-response.dto.ts
@@ -1,0 +1,11 @@
+import { ReceivedMessageResponse } from './received-message-response.dto';
+
+export class MessageListResponse {
+  messages: ReceivedMessageResponse[];
+  pagination: {
+    total: number;
+    limit: number;
+    offset: number;
+    hasMore: boolean;
+  };
+}
