+++ src/messages/dto/sent-message-response.dto.ts
@@ -1,0 +1,15 @@
+export class SentMessageResponse {
+  id: string;
+  platform: string;
+  jobId: string | null;
+  providerMessageId: string | null;
+  targetChatId: string;
+  targetUserId: string | null;
+  targetType: string;
+  messageText: string | null;
+  messageContent: Record<string, unknown> | null;
+  status: string;
+  errorMessage: string | null;
+  sentAt: Date | null;
+  createdAt: Date;
+}
