+++ src/platforms/dto/platform-log-stats-response.dto.ts
@@ -1,0 +1,13 @@
+export class PlatformLogStatsResponse {
+  summary: Array<{
+    level: string;
+    category: string;
+    count: number;
+  }>;
+  recentErrors: Array<{
+    message: string;
+    category: string;
+    timestamp: string;
+    platform: string;
+  }>;
+}
