+++ src/platforms/dto/platform-response.dto.ts
@@ -1,0 +1,9 @@
+export class PlatformResponse {
+  id: string;
+  platform: string;
+  isActive: boolean;
+  testMode: boolean;
+  webhookUrl: string;
+  createdAt: Date;
+  updatedAt: Date;
+}
