+++ src/platforms/dto/query-platform-logs.dto.ts
@@ -1,0 +1,9 @@
+export class QueryPlatformLogsDto {
+  platform?: string;
+  level?: string;
+  category?: string;
+  startDate?: string;
+  endDate?: string;
+  limit?: number;
+  offset?: number;
+}
