+++ src/platforms/dto/supported-platforms-response.dto.ts
@@ -1,0 +1,17 @@
+export class SupportedPlatformsResponse {
+  platforms: Array<{
+    name: string;
+    displayName: string;
+    connectionType: string;
+    features: {
+      supportsWebhooks: boolean;
+      supportsPolling: boolean;
+      supportsWebSocket: boolean;
+    };
+    credentials: {
+      required: string[];
+      optional: string[];
+      example: Record<string, any>;
+    } | null;
+  }>;
+}
