+++ src/platforms/messages/dto/message-retry-response.dto.ts
@@ -1,0 +1,5 @@
+export class MessageRetryResponse {
+  success: boolean;
+  jobId: string;
+  message: string;
+}
