+++ src/platforms/messages/dto/message-send-response.dto.ts
@@ -1,0 +1,13 @@
+export class MessageSendResponse {
+  success: boolean;
+  jobId: string;
+  status: string;
+  targets: Array<{
+    platformId: string;
+    type: string;
+    id: string;
+  }>;
+  platformIds: string[];
+  timestamp: string;
+  message: string;
+}
