+++ src/platforms/messages/dto/message-status-response.dto.ts
@@ -1,0 +1,9 @@
+export class MessageStatusResponse {
+  jobId: string;
+  status: string;
+  progress?: number;
+  result?: any;
+  error?: string;
+  createdAt: Date;
+  updatedAt?: Date;
+}
