+++ src/projects/dto/project-response.dto.ts
@@ -1,0 +1,14 @@
+export class ProjectResponse {
+  id: string;
+  name: string;
+  description?: string;
+  slug: string;
+  environment: 'development' | 'staging' | 'production';
+  isDefault: boolean;
+  settings?: Record<string, unknown>;
+  createdAt: string;
+  updatedAt: string;
+  _count?: {
+    apiKeys: number;
+  };
+}
