+++ src/users/dto/add-member.dto.ts
@@ -1,0 +1,6 @@
+import { ProjectRole } from '../../common/types/enums';
+
+export class AddMemberDto {
+  email: string;
+  role: ProjectRole;
+}
