+++ src/users/dto/project-member-response.dto.ts
@@ -1,0 +1,15 @@
+import { ProjectRole } from '../../common/types/enums';
+
+export class ProjectMemberResponse {
+  id: string;
+  projectId: string;
+  userId: string;
+  role: ProjectRole;
+  createdAt: string;
+  updatedAt: string;
+  user: {
+    id: string;
+    email: string;
+    name?: string;
+  };
+}
