+++ src/users/dto/update-member-role.dto.ts
@@ -1,0 +1,5 @@
+import { ProjectRole } from '../../common/types/enums';
+
+export class UpdateMemberRoleDto {
+  role: ProjectRole;
+}
